# docs/yaml/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce Backend API
  description: This is an OpenAPI Specification for my custom E-Commerce API. Read the following information to know how different API endpoints work.
  version: 1.0.0
  contact:
    name: Alan Ng
    email: nkk991026@gmail.com
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "http://localhost:10888/api/"
    description: Development server
...
paths:
  /users/:
    get:
      tags:
        - Users
      description:
        Get all users
      responses:
        '200':
          description: View all user record
    post:
      tags:
        - Users
      description:
        Create new user
      requestBody:
        description: New user data
        required: true
        content:
          application/json:
            schema:
              $ref: ./users.json#/definitions/user
      responses:
        '201':
          description: User record created.
        '500':
          description: Data create failed.
  /users/{username}:
    get:
      tags:
        - Users
      description:
        Get specific user by Username
      responses:
        '200':
          description: View specific user record
        '404':
          description: User record not found
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the user
    get:
      tags:
        - Users
      description:
        Get specific user by ID
      responses:
        '200':
          description: View specific user record
        '404':
          description: User record not found
    put:
      tags:
        - Users
      description:
        Update specific user by ID
      responses:
        '200':
          description: User record updated
        '500':
          description: User record update failed
    delete:
      tags:
        - Users
      description:
        Delete specific user by ID
      responses:
        '200':
          description: User record updated
        '404':
          description: User record not found
        '500':
          description: User record update failed
  # Products
...